+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [BREAKING CHANGE] Renamed the `cancelOrderWithSigner` method to `cancelOrder`
 - [BREAKING CHANGE] Renamed the `createTradeWithSigner` method to `createTrade`
 
+- Changed the `getAddress` method from the `L2Signer` interface to be able to return both `string` or async `Promise<string>`
+
 ### Removed
 
 - [BREAKING CHANGE] Removed the deprecated `registerOffchain` method